requirements.txt
google_api_python_client==2.116.0
google_auth_oauthlib==1.2.0
pandas==2.0.3
protobuf==3.20.3
psycopg2==2.9.9
streamlit==1.28.1
tabulate==0.8.10
